Migrar importes monetarios a BigDecimal

+++ src/innerjoinsquad/modelo/ClienteEstandar.java
@@ -3,6 +3,8 @@ package innerjoinsquad.modelo;
 import jakarta.persistence.Entity;
 import jakarta.persistence.DiscriminatorValue;
 
+import java.math.BigDecimal;
+
 @Entity
 @DiscriminatorValue("ESTANDAR")
 
@@ -22,7 +24,7 @@ public class ClienteEstandar extends Cliente {
     }
 
     @Override
-    public double aplicarDescuentoEnvio(double gastosEnvioBase) {
+    public BigDecimal aplicarDescuentoEnvio(BigDecimal gastosEnvioBase) {
         return gastosEnvioBase; // sin descuento
     }
